rename RecentSearch class to Search combine SearchValue into Search add filtering parameters to Search class rename Util to DateUtil and remove functions out of the object remove searchEvent function out of parent of object add searchNSave function to search and filter the events list and to add recent searches to room data base table
