+++ app/src/main/java/com/lambda_labs/community_calendar/Repository.kt
@@ -1,10 +1,8 @@
 package com.lambda_labs.community_calendar
 
 import android.content.Context
-import android.os.AsyncTask
-import androidx.lifecycle.LiveData
 import androidx.room.Room.databaseBuilder
-import com.lambda_labs.community_calendar.model.RecentSearch
+import com.lambda_labs.community_calendar.model.Search
 import com.lambda_labs.community_calendar.room.RecentSearchDatabase
 import io.reactivex.Flowable
 
@@ -13,19 +11,19 @@ class Repository(context: Context) {
         databaseBuilder(context, RecentSearchDatabase::class.java, "recent_search_db").build()
     }
 
-    fun getRecentSearchList(): Flowable<MutableList<RecentSearch>>{
+    fun getRecentSearchList(): Flowable<MutableList<Search>>{
         return recentSearchDatabase.dataAccessObject().getRecentSearchList()
     }
 
-    fun addRecentSearch(recentSearch: RecentSearch){
-        recentSearchDatabase.dataAccessObject().addRecentSearch(recentSearch)
+    fun addRecentSearch(search: Search){
+        recentSearchDatabase.dataAccessObject().addRecentSearch(search)
     }
 
-    fun updateRecentSearch(recentSearch: RecentSearch){
-        recentSearchDatabase.dataAccessObject().updateRecentSearch(recentSearch)
+    fun updateRecentSearch(search: Search){
+        recentSearchDatabase.dataAccessObject().updateRecentSearch(search)
     }
 
-    fun removeRecentSearch(recentSearch: RecentSearch){
-        recentSearchDatabase.dataAccessObject().removeRecentSearch(recentSearch)
+    fun removeRecentSearch(search: Search){
+        recentSearchDatabase.dataAccessObject().removeRecentSearch(search)
     }
 }
