+++ app/src/main/java/com/lambda_labs/community_calendar/SearchFragment.kt
@@ -2,18 +2,18 @@ package com.lambda_labs.community_calendar
 
 import android.content.Context
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import android.widget.TextView
 import androidx.constraintlayout.widget.ConstraintSet
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.lambda_labs.community_calendar.model.RecentSearch
+import com.lambda_labs.community_calendar.model.Search
 import com.lambda_labs.community_calendar.util.hideKeyboard
 import com.lambda_labs.community_calendar.viewmodel.SearchViewModel
 import kotlinx.android.synthetic.main.activity_main.*
@@ -50,7 +50,7 @@ class SearchFragment : Fragment() {
         searches_recycler.layoutManager = LinearLayoutManager(activity, RecyclerView.VERTICAL, false)
 
         viewModel.getRecentSearches()
-        viewModel.searchList.observe(viewLifecycleOwner, Observer<MutableList<RecentSearch>> { recentSearches ->
+        viewModel.searchList.observe(viewLifecycleOwner, Observer<MutableList<Search>> { recentSearches ->
             searches_recycler.adapter = RecentSearchRecycler(recentSearches)
         })
     }
@@ -70,7 +70,7 @@ class SearchFragment : Fragment() {
         super.onDestroy()
     }
 
-    inner class RecentSearchRecycler(private val recentSearches: MutableList<RecentSearch>) :
+    inner class RecentSearchRecycler(private val searches: MutableList<Search>) :
         RecyclerView.Adapter<RecentSearchRecycler.ViewHolder>() {
 
         override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -79,10 +79,10 @@ class SearchFragment : Fragment() {
             return ViewHolder(view)
         }
 
-        override fun getItemCount(): Int = recentSearches.size
+        override fun getItemCount(): Int = searches.size
 
         override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-            val recentSearch = recentSearches[position]
+            val recentSearch = searches[position]
 
             holder.searchText.text = recentSearch.searchText
         }
