+++ app/src/main/java/com/lambda_labs/community_calendar/room/RecentSearchDao.kt
@@ -1,20 +1,20 @@
 package com.lambda_labs.community_calendar.room
 
 import androidx.room.*
-import com.lambda_labs.community_calendar.model.RecentSearch
+import com.lambda_labs.community_calendar.model.Search
 import io.reactivex.Flowable
 
 @Dao
 interface RecentSearchDao {
-    @Query("SELECT * FROM RecentSearch")
-    fun getRecentSearchList(): Flowable<MutableList<RecentSearch>>
+    @Query("SELECT * FROM Search")
+    fun getRecentSearchList(): Flowable<MutableList<Search>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    fun addRecentSearch(recentSearch: RecentSearch)
+    fun addRecentSearch(search: Search)
 
     @Update(onConflict = OnConflictStrategy.REPLACE)
-    fun updateRecentSearch(recentSearch: RecentSearch)
+    fun updateRecentSearch(search: Search)
 
     @Delete
-    fun removeRecentSearch(recentSearch: RecentSearch)
+    fun removeRecentSearch(search: Search)
 }
