+++ app/src/main/java/com/lambda_labs/community_calendar/room/RecentSearchDatabase.kt
@@ -2,9 +2,9 @@ package com.lambda_labs.community_calendar.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.lambda_labs.community_calendar.model.RecentSearch
+import com.lambda_labs.community_calendar.model.Search
 
-@Database(entities = [RecentSearch::class], version = 1, exportSchema = false)
+@Database(entities = [Search::class], version = 1, exportSchema = false)
 abstract class RecentSearchDatabase: RoomDatabase() {
     abstract fun dataAccessObject(): RecentSearchDao
 }
