+++ app/src/main/java/com/lambda_labs/community_calendar/viewmodel/SearchViewModel.kt
@@ -4,8 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.lambda_labs.community_calendar.App
-import com.lambda_labs.community_calendar.model.RecentSearch
-import io.reactivex.Flowable
+import com.lambda_labs.community_calendar.model.Search
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.schedulers.Schedulers
@@ -13,25 +12,25 @@ import io.reactivex.schedulers.Schedulers
 class SearchViewModel: ViewModel() {
     // Database call will be done in viewmodel
     private var disposable: Disposable? = null
-    private val recentSearchList: MutableLiveData<MutableList<RecentSearch>> = MutableLiveData(
+    private val recentSearchList: MutableLiveData<MutableList<Search>> = MutableLiveData(
         mutableListOf())
-    val searchList: LiveData<MutableList<RecentSearch>> = recentSearchList
+    val searchList: LiveData<MutableList<Search>> = recentSearchList
 
     // Retrieves searches stored in database and saves them to recentSearchList to pass to searchList for fragment
     fun getRecentSearches(){
         disposable = App.repository.getRecentSearchList().subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread()).subscribe{ recentSearches: MutableList<RecentSearch> ->
+            .observeOn(AndroidSchedulers.mainThread()).subscribe{ searches: MutableList<Search> ->
                 recentSearchList.value?.clear()
-                recentSearchList.value?.addAll(recentSearches)
+                recentSearchList.value?.addAll(searches)
             }
     }
 
-    fun updateRecentSearch(recentSearch: RecentSearch){
-        App.repository.updateRecentSearch(recentSearch)
+    fun updateRecentSearch(search: Search){
+        App.repository.updateRecentSearch(search)
     }
 
-    fun removeRecentSearch(recentSearch: RecentSearch){
-        App.repository.removeRecentSearch(recentSearch)
+    fun removeRecentSearch(search: Search){
+        App.repository.removeRecentSearch(search)
     }
 
     override fun onCleared() {
